gems baseCost to Double

+++ src/main/java/objects/resources/Gems.java
@@ -3,7 +3,7 @@ package objects.resources;
 import objects.abstracts.AbstractResource;
 
 public class Gems extends AbstractResource {
-    public Gems(String name, String description, Integer baseCost, Integer size) {
+    public Gems(String name, String description, Double baseCost, Integer size) {
         super("gems", description, baseCost, size);
     }
 }
